+++ FM2KHook/src/simple_css_sync.h
@@ -11,11 +11,11 @@ struct CSSMessage {
     uint8_t p1_cursor_x;          // P1 cursor X (low byte of u32)
     uint8_t p1_cursor_y;          // P1 cursor Y (low byte of u32)
     uint8_t p1_confirmed;         // P1 confirmation status
-    uint8_t p1_color_button;      // P1 color button (0x00, 0x20, 0x40, 0x80, 0x100, 0x200)
+    uint16_t p1_color_button;     // P1 color button (0x010, 0x020, 0x040, 0x080, 0x100, 0x200)
     uint8_t p2_cursor_x;          // P2 cursor X (low byte of u32)
     uint8_t p2_cursor_y;          // P2 cursor Y (low byte of u32)
     uint8_t p2_confirmed;         // P2 confirmation status
-    uint8_t p2_color_button;      // P2 color button (0x00, 0x20, 0x40, 0x80, 0x100, 0x200)
+    uint16_t p2_color_button;     // P2 color button (0x010, 0x020, 0x040, 0x080, 0x100, 0x200)
 };
 
 class SimpleCSSSync {
@@ -38,7 +38,7 @@ public:
     bool Initialize(bool is_host, uint16_t base_port, const char* remote_ip = "127.0.0.1");
     
     // Update local CSS state
-    void UpdateLocalState(uint8_t cursor_x, uint8_t cursor_y, uint8_t confirmed, uint8_t color_button = 0);
+    void UpdateLocalState(uint8_t cursor_x, uint8_t cursor_y, uint8_t confirmed, uint16_t color_button = 0);
     
     // Send/receive CSS updates (non-blocking)
     bool SendUpdate();
